+++ src/models/animations/Animation.cpp
@@ -40,8 +40,8 @@ namespace framework {
 
 Animation::Animation(aiAnimation *assimpAnimation) {
   mName = assimpAnimation->mName.C_Str();
-  mDuration = assimpAnimation->mDuration;
-  mTicksPerSec = assimpAnimation->mTicksPerSecond;
+  mDuration = static_cast<float>(assimpAnimation->mDuration);
+  mTicksPerSec = static_cast<float>(assimpAnimation->mTicksPerSecond);
   mChannels.reserve(assimpAnimation->mNumChannels);
   for (unsigned int i = 0; i < assimpAnimation->mNumChannels; i++) {
     mChannels.push_back(new Channel(assimpAnimation->mChannels[i]));
